requirements.txt
certifi==2024.8.30
charset-normalizer==3.4.0
click==8.1.7
Flask==2.1.0
Flask-Cors==4.0.1
Flask-JWT-Extended==4.6.0
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
itsdangerous==2.2.0
jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
PyJWT==2.9.0
python-dotenv==1.0.1
requests==2.32.3
SQLAlchemy==1.4.22
urllib3==2.2.3
Werkzeug==2.1.1
zipp==3.20.2
